requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
pymysql>=1.0.0
python-agi>=0.9.0
pyyaml>=6.0
requests>=2.28.0
ratelimit>=2.2.1
cachetools>=5.3.0
python-json-logger>=2.0.0
redis>=4.5.0

# Optional for migrations
alembic>=1.12.0

# Monitoring
prometheus-client>=0.17.0

# Additional security/encryption
cryptography>=3.4.7

# Circuit breaker
pybreaker>=0.6.0

# (Optional) Anomaly detection – adjust as needed
anomaly_detector>=1.0.0
